Create SOLID classes for person creation
