+++ list_people.rb
@@ -1,0 +1,8 @@
+class ListPeople
+  def list_people(people)
+    @people = people
+    puts 'People list'
+    puts 'No people added yet' if @people.empty?
+    @people.each { |person| puts person.name }
+  end
+end
